backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Database - MongoDB Only
pymongo==4.6.1
motor==3.3.2
certifi==2024.2.2
dnspython==2.4.2

# Neo4j Graph Database
neo4j==5.14.1

# AI & ML (Python 3.12 compatible versions)
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5
langgraph==0.0.20
sentence-transformers==2.3.1
faiss-cpu==1.8.0
openai==1.10.0

# Document Processing for RAG
PyPDF2==3.0.1
python-docx==1.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.4.0

# HTTP & APIs
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.1

# Environment
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pyyaml==6.0.1
